+++ src/app/document/services/index.ts
@@ -1,3 +1,2 @@
 export { AuthService } from './auth.service';
 export { DocumentService } from './document.service';
-export { FileSaverService } from './file-saver.service';
