+++ src/app/shared/navbar/navbar.component.html
@@ -1,5 +1,5 @@
 <nav class="docs-navbar mat-elevation-z6">
-  <a mat-button class="docs-button" routerLink="/">
+  <a mat-button class="docs-button" href="/">
     <img class="docs-angular-logo" [src]="logo"
       [alt]="product">
       <span>{{product}}</span>
