+++ src/app/shared/services/index.ts
@@ -1,1 +1,2 @@
 export { ApiService } from './api.service';
+export { FileSaverService } from './file-saver.service';
